client: fix app id size
The size of the type representing the id of a controlled app was the
wrong size: 32 bits instead of 64 like the protocol mandates (see:
proto/protocol.proto:34).

The bug hasn't been observed during any test, but was caught during code
review.

+++ client/applist.h
@@ -17,7 +17,7 @@ namespace msgs {
 class App {
     // TODO: extract icon data
 public:
-    typedef quint32 Id;
+    typedef quint64 Id;
     enum {
         INVALID_ID = (Id) -1
     };
